Upgraded to a later k3po version to resolve some test failures. Added test cases for request.fragmented.with.content.length.
