+++ pom.xml
@@ -45,7 +45,7 @@
 
     <junit.version>4.12</junit.version>
     <jmock.version>2.6.0</jmock.version>
-    <k3po.version>3.0.0-alpha-54</k3po.version>
+    <k3po.version>3.0.0-alpha-63</k3po.version>
     <jmh.version>1.12</jmh.version>
     <nukleus.plugin.version>0.7.3</nukleus.plugin.version>
     <nukleus.version>0.2</nukleus.version>
